+++ ReleaseNotes.md
@@ -1,6 +1,8 @@
 ## 0.0.5
 
-+ taking advantages of the latest SOAR release (1.1.5) for performance improvement.
++ Existing SQL tables can be automatically converted to become **newsql** compatible. This can be turned off by _config()_.
+
++ taking advantages of the latest SOAR release (1.1.6) for performance improvement.
 
 + changing the signature of the _execute()_ function so it's easier to reuse SQL templates. The old format still works, but is deprecated.